pyproject.toml
[project]
name = "form-app"
version = "0.1.0"
description = "Senior Developer Exercise"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.7",
]

[tool.ruff]
target-version = "py314"